0Dev Reference Docs/TRA Files for Reference/requirements.txt
flask
python-dotenv
requests
Pillow
staticmap
werkzeug
urllib3
flask-pymongo
pymongo
flask-login
flask-bcrypt
flask-limiter
geopy
celery
redis
